requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.0
pydantic==1.10.9
python-multipart==0.0.9
httpx==0.27.2

# Core utilities
PyYAML==6.0.1

# Rasa integration (for chatbot proxy/testing)
rasa==3.6.20

# MongoDB stack
pymongo==3.13.0
mongoengine==0.24.2
motor==3.3.2

# OCR stack (optional but supported)
paddleocr==2.7.0.3
opencv-python-headless==4.10.0.84
pdf2image==1.17.0
pypdf==4.3.1

# QR / imaging
qrcode==7.4.2
Pillow==10.0.1

# Auth & Security
PyJWT==2.10.1
bcrypt==4.2.1

# ML
joblib==1.3.2
pandas==2.2.2
scikit-learn==1.3.2

# Google Drive API client (optional)
google-api-python-client==2.97.0
google-auth==2.23.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0

# OpenAI client (optional, for future use)
openai==1.54.3
